Split encryption backend
Separate OpenSSL and QAESEncryption as two backends for MCrypto allowing
for simpler library dependecies.
